fix(scully-plugin-puppeteer): fixed exit problem with ppt renderer
