+++ scully.heavy.config.ts
@@ -1,8 +1,13 @@
-import { config as original } from './scully.sample-blog.config';
+import { ScullyConfig } from 'libs/scully/src/index.js';
+import { cpus } from 'os';
+import { config as original } from './scully/scully.sample-blog.config.js';
 
-export const config = Object.assign({}, original);
+export const config: Promise<Partial<ScullyConfig>> = (async () => {
+  const config = Object.assign({}, await original);
 
-config.routes['/demo/:id'].numberOfPages = 1500;
-config.routes['/user/:userId'].userId.resultsHandler = undefined;
-/** crank it up a notch, more threads seems more likely to fail. */
-config.maxRenderThreads = 32;
+  config.routes['/demo/:id'].numberOfPages = 5000;
+  config.routes['/user/:userId'].userId.resultsHandler = undefined;
+  /** crank it up a notch, more threads seems more likely to fail. */
+  config.maxRenderThreads = cpus().length * 4;
+  return config;
+})();
